add revit sample files in hbJSON format
This commit adds 3 new models for revit sample files. The files are the basic and advanced sample models.

You will see the the advanced model doesn't load correctly because the points are not sorted to follow what the viewer expect. There is a 3rd file (2020_rac_advanced_sample_project_sorted_points.json) which is the same sample  model with sorted points. Sorting the points solves most of the problems for visualization.